Memory fix (#5)
Added a parameter (calc_dist_matrix) in the createSTmodel function which controls whether the square distance matrix is calculated.  By default this is TRUE for backwards comparability and because you want the distance matrix for model fitting.  When createSTmodel is called in predict function to cast unobserved data into STmodel object, the square matrix will no longer be calculated as this is unnecessary and takes a lot of time and memory.